@@ -6,7 +6,7 @@
     "dependencies": {
         "@emotion/react": "^11.8.2",
         "@emotion/styled": "^11.8.1",
-        "@mui/icons-material": "^5.5.1",
+        "@mui/icons-material": "^5.11.0",
         "@mui/material": "^5.5.2",
         "@mui/x-data-grid": "^5.7.0",
         "@mui/x-date-pickers": "^5.0.0-beta.2",
@@ -21,24 +21,29 @@
         "dotenv": "^16.0.1",
         "material-ui-confirm": "^3.0.5",
         "mui-image": "^1.0.4",
+        "npm": "^9.2.0",
         "parameterize": "^1.0.0",
-        "react": "^17.0.2",
+        "react": "^18.2.0",
         "react-burger-menu": "^3.0.6",
         "react-chartjs-2": "^4.0.1",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.2.0",
         "react-dropzone": "^12.0.4",
-        "react-facebook": "^8.1.4",
-        "react-facebook-login": "^4.1.1",
-        "react-google-login": "^5.2.2",
         "react-html-parser": "^2.0.2",
         "react-image-size": "^1.0.5",
+        "react-redux": "^8.0.5",
         "react-router-dom": "^6.2.2",
         "react-scripts": "5.0.0",
         "recharts": "^2.1.9",
+        "redux": "^4.2.0",
+        "react-facebook": "^8.1.4",
+        "react-facebook-login": "^4.1.1",
+        "react-google-login": "^5.2.2",
+        "redux-persist": "^6.0.0",
         "rehype-parse": "^8.0.4",
         "rehype-stringify": "^9.0.3",
         "unified": "^10.1.2",
         "unist-util-visit": "^4.1.0",
+        "update": "^0.7.4",
         "web-vitals": "^2.1.4"
     },
     "scripts": {
