@@ -1,20 +1,33 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import Context from './store/Context'
 import reportWebVitals from './reportWebVitals';
 import CssBaseline from '@mui/material/CssBaseline';
 
-ReactDOM.render(
+// Redux Setup 
+import { Provider } from 'react-redux'
+import { PersistGate } from 'redux-persist/integration/react' // for persisting data
+import { store, persistor } from './store/store'
+
+// subscribe 
+
+// subscribe this function is used for console the current state if there is no Redux extension there
+store.subscribe(() => console.log(store.getState()));
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(
   <React.StrictMode>
-    <Context>
-    <CssBaseline/>
-    <App />
-    </Context>
-  </React.StrictMode>,
-  document.getElementById('root')
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistor}>
+          <CssBaseline />
+          <App />
+      </PersistGate>
+    </Provider>
+  </React.StrictMode>
 );
+
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(//console.log))
