@@ -1,15 +1,13 @@
 import React, { useState, useEffect } from "react";
 import {
-  // Tabs,
-  // Tab,
   IconButton,
   Menu,
   MenuItem,
   Box,
   Grid,
   Button,
-  ListItemIcon,
   List,
+  ListItemIcon,
   ListItem,
   ListItemText,
   Typography,
@@ -21,8 +19,6 @@ import {
 
 } from "@mui/material";
 import "../assets/custom/css/home.css";
-// import Slide from "@mui/material/Slide";
-// import Backdrop from "@mui/material/Backdrop";
 import logo from "../assets/img/Blog/logo.webp";
 
 // icons
@@ -35,12 +31,8 @@ import WbSunnyIcon from "@mui/icons-material/WbSunny";
 import DarkModeIcon from "@mui/icons-material/DarkMode";
 import ShoppingBagOutlinedIcon from "@mui/icons-material/ShoppingBagOutlined";
 import PeopleAltOutlinedIcon from "@mui/icons-material/PeopleAltOutlined";
-// import ExploreOutlinedIcon from "@mui/icons-material/ExploreOutlined";
-// import CardGiftcardOutlinedIcon from "@mui/icons-material/CardGiftcardOutlined";
 import AddIcon from '@mui/icons-material/Add';
 import EmojiEventsIcon from '@mui/icons-material/EmojiEvents';
-// import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-// import PersonOutlineOutlinedIcon from "@mui/icons-material/PersonOutlineOutlined";
 import CategoryIcon from '@mui/icons-material/Category';
 import InboxIcon from '@mui/icons-material/Inbox';
 import SettingsOutlinedIcon from "@mui/icons-material/SettingsOutlined";
@@ -49,7 +41,6 @@ import CloseIcon from "@mui/icons-material/Close";
 import LogoutIcon from "@mui/icons-material/Logout";
 import ViewCarouselIcon from "@mui/icons-material/ViewCarousel";
 import AdminPanelSettingsIcon from '@mui/icons-material/AdminPanelSettings';
-// import CollectionsIcon from "@mui/icons-material/Collections";
 import DraftsIcon from '@mui/icons-material/Drafts';
 import GridViewIcon from '@mui/icons-material/GridView';
 import InventoryIcon from '@mui/icons-material/Inventory';
@@ -57,30 +48,28 @@ import MergeIcon from '@mui/icons-material/Merge';
 import HardwareIcon from '@mui/icons-material/Hardware';
 import AssignmentIndIcon from '@mui/icons-material/AssignmentInd';
 import AccountTreeIcon from '@mui/icons-material/AccountTree';
-// import state 
-import { Store } from '../store/Context'
-import { DarkMode, SideTabs, Auth } from '../store/Types'
+
+// import redux 
+import {setAlert, setMode,setTab} from '../store/action/action'
+import { useDispatch,useSelector } from "react-redux";
 
 const Home = (props) => {
 
-  const { state, dispatch } = Store()
+  const dispatch = useDispatch()
+  const {auth,mode,tab} = useSelector(state=>state);
 
   const history = props.history;
 
 
   // states
   const [anchor, setAnchor] = useState(null);
-  const [type, setType] = useState({
-    product: false,
-    admin: false,
-    order : false
-  });
+ 
 
   useEffect(() => {
     // console.log(state.Auth.isLogin)
-    if (state.Auth.isLogin === false)
+    if (auth.isLogin === false)
       history("/");
-  }, [state.Auth.isLogin]);
+  }, [auth.isLogin]);
 
   const handleMenuClose = () => {
     setAnchor(null);
@@ -90,204 +79,25 @@ const Home = (props) => {
     setAnchor(e.currentTarget);
   };
 
-  const handleClose = () => {
-    dispatch({
-      type: SideTabs,
-      payload: {
+  const handleClose = (direct = undefined) => {
+    handleMenuClose();
+    if(direct) history(direct);
+    return dispatch(setTab({
         open: false
       }
-    })
-    handleMenuClose();
+    ))
   };
-
-
-
-  //   return {
-  //     id: `vertical-tab-${index}`,
-  //     "aria-controls": `vertical-tabpanel-${index}`,
-  //   };
-  // }
-
-// this tab commented because client demand something new and compact to show   
-  // function VerticalTabs() {
-
-  //   const handleChange = (event, newValue) => {
-  //     history(`${ModuleName[newValue]}`)
-  //     setValue(newValue);
-  //   };
-
-  //   return (
-  //     <Box sx={{ flexGrow: 1, display: "flex", width: "100%" }}>
-  //       {state.SideTabs.open && (
-  //         <Slide direction="right" in={state.SideTabs.open} mountOnEnter unmountOnExit>
-  //           <Backdrop
-  //             sx={{ color: "#fff", zIndex: (theme) => theme.zIndex.drawer + 1 }}
-  //             open={state.SideTabs.open}
-  //             onClick={handleClose}
-  //           >
-  //             <Tabs
-  //               orientation="vertical"
-  //               value={value}
-  //               variant="scrollable"
-  //               onChange={handleChange}
-  //               indicatorColor="primary"
-  //               textColor="primary"
-  //               aria-label="Vertical tabs example"
-  //               className={state.DarkMode.mode === true ? "darkTabs2" : "tabs2"}
-  //               sx={{
-  //                 borderRight: 1,
-  //                 borderColor: "divider",
-  //                 ".MuiTabs-indicator": {
-  //                   left: 0,
-  //                   width: "5px",
-  //                 },
-  //               }}
-  //             >
-
-
-  //               <Tab
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<GridViewIcon />}
-  //                 label="DashBoard"
-  //                 {...a11yProps(0)}
-  //               />
-
-  //               <Accordion>
-  //                 <AccordionSummary
-  //                   expandIcon={<ExpandMoreIcon />}
-  //                   aria-controls="panel1a-content"
-  //                   id="panel1a-header"
-  //                 >
-  //                   <Typography>Accordion 1</Typography>
-  //                 </AccordionSummary>
-  //                 <AccordionDetails>
-  //                   <Typography>
-  //                     Product
-  //                   </Typography>
-  //                 </AccordionDetails>
-  //               </Accordion>
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<ShoppingBagOutlinedIcon />}
-  //                 label="Product"
-  //                 {...a11yProps(1)}
-  //               />
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<MergeIcon />}
-  //                 label="Merge Product"
-  //                 {...a11yProps(2)}
-  //               />
-
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<CollectionsIcon />}
-  //                 label="Gallery"
-  //                 {...a11yProps(3)}
-  //               />
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<ViewCarouselIcon />}
-  //                 label="Banner"
-  //                 {...a11yProps(4)}
-  //               />
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<PeopleAltOutlinedIcon />}
-  //                 label="Customer"
-  //                 {...a11yProps(5)}
-  //               />
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<ExploreOutlinedIcon />}
-  //                 label="Order"
-  //                 {...a11yProps(6)}
-  //               />
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<CardGiftcardOutlinedIcon />}
-  //                 label="Coupons"
-  //                 {...a11yProps(7)}
-  //               />
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<ArticleIcon />}
-  //                 label="Blog"
-  //                 {...a11yProps(8)}
-  //               />
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<AdminPanelSettingsIcon />}
-  //                 label="Hardware"
-  //                 {...a11yProps(9)}
-  //               />
-  //               {localStorage.getItem('role') === 'Super Admin' && <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<DraftsIcon />}
-  //                 label="Draft"
-  //                 {...a11yProps(10)}
-  //               />}
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<InventoryIcon />}
-  //                 label="Stock Channel"
-  //                 {...a11yProps(11)}
-  //               />
-  //               <Tab
-
-  //                 sx={{ justifyContent: 'left !important' }}
-  //                 iconPosition="start"
-  //                 icon={<SettingsOutlinedIcon />}
-  //                 label="Profile"
-  //                 {...a11yProps(12)}
-  //               />
-  //               <Button
-  //                 color="primary"
-  //                 sx={{ margin: "auto" }}
-  //                 startIcon={<LogoutIcon />}
-  //                 variant="contained"
-  //               >
-  //                 Log Out
-  //               </Button>
-  //             </Tabs>
-  //           </Backdrop>
-  //         </Slide>
-  //       )}
-
-
-
-  //     </Box>
-  //   );
-  // }
 
 
   // side bar for the tabs 
   function SwipeableTemporaryDrawer() {
+    const [type, setType] = useState({
+      product: false,
+      admin: false,
+      order : false
+    });
 
-    const list = () => (
+    function ListTabs(){return (
       <Box
         sx={{ flexGrow: 1, display: "flex", width: 250, flexDirection: 'column' }}
         role="presentation"
@@ -309,7 +119,7 @@ const Home = (props) => {
               bgcolor: 'background.paper',
             }} component="nav" aria-label="mailbox folders">
 
-              <ListItem button onClick={() => { history('/dashboard'); handleClose(); }}>
+              <ListItem  onClick={() => { handleClose('/dashboard'); }}>
                 <ListItemAvatar >
                   <Avatar
                     sx={{
@@ -345,7 +155,7 @@ const Home = (props) => {
               <Collapse in={type.product} timeout="auto" unmountOnExit>
                 <List dense = {true} component="div" disablePadding>
 
-                  <ListItem sx={{ pl: 4 }} button onClick={() => { history('/products'); handleClose(); }}>
+                  <ListItem sx={{ pl: 4 }}  onClick={() => {  handleClose('/products'); }}>
                     <ListItemAvatar>
                       <Avatar
                         sx={{
@@ -360,7 +170,7 @@ const Home = (props) => {
                     </ListItemAvatar>
                     <ListItemText primary="Product" />
                   </ListItem>
-                  <ListItem button sx={{ pl: 4 }} onClick={() => { history('/merge'); handleClose(); }}>
+                  <ListItem  sx={{ pl: 4 }} onClick={() => { handleClose('/merge'); }}>
                     <ListItemAvatar>
                       <Avatar
                         sx={{
@@ -376,7 +186,7 @@ const Home = (props) => {
                     <ListItemText primary="Merge Product" />
                   </ListItem >
 
-                  <ListItem button sx={{ pl: 4 }} onClick={() => { history('/variation'); handleClose(); }}>
+                  <ListItem  sx={{ pl: 4 }} onClick={() => { handleClose('/variation'); }}>
                     <ListItemAvatar>
                       <Avatar
                         sx={{
@@ -394,7 +204,7 @@ const Home = (props) => {
                 </List>
               </Collapse>
 
-              <ListItem button onClick={() => { history('/action'); handleClose(); }}>
+              <ListItem  onClick={() => {  handleClose('/action'); }}>
                 <ListItemAvatar >
                   <Avatar
                     sx={{
@@ -411,7 +221,7 @@ const Home = (props) => {
               </ListItem>
 
 
-              <ListItem button onClick={(e) => { setType({ ...type, order: !type.order }) }}>
+              <ListItem  onClick={(e) => { setType({ ...type, order: !type.order }) }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -432,7 +242,7 @@ const Home = (props) => {
               <Collapse in={type.order} timeout="auto" unmountOnExit>
                 <List dense = {true} component="div" disablePadding>
 
-                  <ListItem sx={{ pl: 4 }} button  onClick={() => { history('/order'); handleClose(); }}>
+                  <ListItem sx={{ pl: 4 }}   onClick={() => {  handleClose('/order'); }}>
                     <ListItemAvatar>
                       <Avatar
                         sx={{
@@ -447,7 +257,7 @@ const Home = (props) => {
                     </ListItemAvatar>
                     <ListItemText primary="Orders" />
                   </ListItem>
-                  <ListItem button sx={{ pl: 4 }} onClick={() => { history('/create_order'); handleClose(); }}>
+                  <ListItem  sx={{ pl: 4 }} onClick={() => { handleClose('/create_order'); }}>
                     <ListItemAvatar>
                       <Avatar
                         sx={{
@@ -465,7 +275,7 @@ const Home = (props) => {
                 </List>
               </Collapse>
 
-              <ListItem button onClick={() => { history('/customer'); handleClose(); }}>
+              <ListItem  onClick={() => { handleClose('/customer'); }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -481,7 +291,7 @@ const Home = (props) => {
                 <ListItemText primary="Customer" />
               </ListItem>
 
-              <ListItem button onClick={() => { history('/reward'); handleClose(); }}>
+              <ListItem  onClick={() => { handleClose('/reward'); }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -497,7 +307,7 @@ const Home = (props) => {
                 <ListItemText primary="Reward" />
               </ListItem>
 
-              <ListItem button onClick={() => { history('/blogs'); handleClose(); }}>
+              <ListItem  onClick={() => {  handleClose('/blogs'); }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -513,7 +323,7 @@ const Home = (props) => {
                 <ListItemText primary="Blog" />
               </ListItem>
 
-              <ListItem button onClick={() => { history('/banner'); handleClose(); }}>
+              <ListItem  onClick={() => {  handleClose('/banner'); }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -529,7 +339,7 @@ const Home = (props) => {
                 <ListItemText primary="Banner" />
               </ListItem>
 
-              <ListItem button onClick={() => { history('/inventory'); handleClose(); }}>
+              <ListItem  onClick={() => { handleClose('/inventory'); }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -546,7 +356,7 @@ const Home = (props) => {
               </ListItem>
 
 
-              <ListItem button onClick={(e) => { setType({ ...type, admin: !type.admin }) }}>
+              <ListItem  onClick={(e) => { setType({ ...type, admin: !type.admin }) }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -567,7 +377,7 @@ const Home = (props) => {
               <Collapse in={type.admin} timeout="auto" unmountOnExit>
                 <List dense = {true} component="div" disablePadding>
 
-                  <ListItem sx={{ pl: 4 }} button onClick={() => { history('/user'); handleClose(); }}>
+                  <ListItem sx={{ pl: 4 }}  onClick={() => {  handleClose('/user'); }}>
                     <ListItemAvatar>
                       <Avatar
                         sx={{
@@ -583,7 +393,7 @@ const Home = (props) => {
                     <ListItemText primary="User" />
                   </ListItem>
 
-                  <ListItem sx={{ pl: 4 }} button onClick={() => { history('/suppliers'); handleClose(); }}>
+                  <ListItem sx={{ pl: 4 }}  onClick={() => {  handleClose('/suppliers'); }}>
                     <ListItemAvatar>
                       <Avatar
                         sx={{
@@ -600,7 +410,7 @@ const Home = (props) => {
                   </ListItem>
 
 
-                  <ListItem button sx={{ pl: 4 }} onClick={() => { history('/settings'); handleClose(); }}>
+                  <ListItem  sx={{ pl: 4 }} onClick={() => {  handleClose('/settings'); }}>
                     <ListItemAvatar>
                       <Avatar
                         sx={{
@@ -618,7 +428,7 @@ const Home = (props) => {
                 </List>
               </Collapse>
 
-              <ListItem button onClick={() => { history('/hardware'); handleClose(); }}>
+              <ListItem  onClick={() => {  handleClose('/hardware'); }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -635,7 +445,7 @@ const Home = (props) => {
               </ListItem>
 
 
-              <ListItem button onClick={() => { history('/accessories'); handleClose(); }}>
+              <ListItem  onClick={() => {  handleClose('/accessories'); }}>
                 <ListItemAvatar>
                   <Avatar
                     sx={{
@@ -669,22 +479,22 @@ const Home = (props) => {
           </Grid>
         </Grid>
       </Box>
-    );
+    )}
 
     return (
       <div>
         <Drawer
           anchor={'left'}
-          open={state.SideTabs.open}
+          open={tab.open}
           onClose={handleClose}
         >
-          {list()}
+          {ListTabs()}
         </Drawer>
       </div>
     );
   }
 
-  function MenuBox(type) {
+  function MenuBox() {
     return (
       <Menu
         id={"menu"}
@@ -709,13 +519,11 @@ const Home = (props) => {
 
   const handleLogout = () => {
     localStorage.clear();
-    dispatch({
-      type: Auth, payload: {
+    dispatch(setAlert({
         isLogin: false,
         WDToken: null,
         role: null
-      }
-    })
+    }))
     handleMenuClose()
     handleClose();
     history("/");
@@ -745,21 +553,18 @@ const Home = (props) => {
         container
         p={1}
         spacing={2}
-        className={state.DarkMode.mode === true ? "darkNav" : "topNav"}
+        className={mode.type === true ? "darkNav" : "topNav"}
         sx={{ boxShadow: 1 }}
       >
         <Grid item xs={4} sx={{ display: "flex" }}>
-          {state.SideTabs.open === false ? (
+          {tab.open === false ? (
             <IconButton
               className="hamIcon"
               onClick={() => {
-                localStorage.setItem("mode", false);
-                dispatch({
-                  type: SideTabs,
-                  payload: {
+                dispatch(setTab({
                     open: true
                   }
-                })
+                ))
               }}
               size="small"
               color="primary"
@@ -770,13 +575,9 @@ const Home = (props) => {
             <IconButton
               className="hamIcon"
               onClick={() => {
-                localStorage.setItem("mode", true);
-                dispatch({
-                  type: SideTabs,
-                  payload: {
+                dispatch(setTab({
                     open: false
-                  }
-                })
+                }))
               }}
               size="small"
               color="primary"
@@ -793,15 +594,13 @@ const Home = (props) => {
         </Grid>
 
         <Grid item sx={{ display: "flex", justifyContent: "end" }} xs={4}>
-          {state.DarkMode.mode === true ? (
+          {mode.type === true ? (
             <IconButton
               onClick={() => {
-                dispatch({
-                  type: DarkMode,
-                  payload: {
-                    mode: false
+                dispatch(setMode({
+                    type: false
                   }
-                });
+                ));
               }}
               size="small"
               color="primary"
@@ -811,12 +610,10 @@ const Home = (props) => {
           ) : (
             <IconButton
               onClick={() => {
-                dispatch({
-                  type: DarkMode,
-                  payload: {
-                    mode: true
-                  }
-                });
+                dispatch(setMode({
+                  type: true
+                }
+              ));
               }}
               size="small"
               color="primary"
@@ -829,9 +626,9 @@ const Home = (props) => {
             <NotificationsIcon />
           </IconButton>
 
-          {MenuBox(type)}
+          {MenuBox()}
 
-          <IconButton onClick={(e) => { handleMenu(e); setType('profile') }} size="small" color="primary">
+          <IconButton onClick={handleMenu} size="small" color="primary">
             <PersonIcon />
           </IconButton>
         </Grid>
@@ -840,7 +637,7 @@ const Home = (props) => {
 
       {/* Sidenav  */}
       {/* {VerticalTabs()} */}
-      {SwipeableTemporaryDrawer()}
+      <SwipeableTemporaryDrawer/>
       {/* Sidenav Ends  */}
     </Box>
   );
