@@ -1,13 +1,14 @@
 import React from 'react';
 import Snackbar from '@mui/material/Snackbar';
 import MuiAlert from '@mui/material/Alert';
-import {Store} from '../store/Context'
-import {Notify} from '../store/Types'
 
+import {useDispatch,useSelector} from  'react-redux'
+import {setAlert} from  '../store/action/action'
 
 export default function SnackBar() {
 
-  const {state,dispatch} = Store()
+  const {alert} = useSelector(state=>state);
+  const dispatch = useDispatch();
 
     const Alert = React.forwardRef(function Alert(props, ref) {
         return <MuiAlert elevation={6} ref={ref} variant="filled" {...props} />;
@@ -15,11 +16,11 @@ export default function SnackBar() {
      
   return (
       <>
-          {state.Notify.open === true  &&   
-          <Snackbar open={state.Notify.open} autoHideDuration={6000} 
-          onClose={()=>dispatch({type : Notify, payload : {open: false, variant : null,massage : null}})}>
-                <Alert onClose={()=>dispatch({type : Notify, payload : {open: false, variant : null,massage : null}})} severity = {state.Notify.variant}  sx={{ width: '100%' }}>
-                    {state.Notify.message}
+          {alert.open === true  &&   
+          <Snackbar open={alert.open} autoHideDuration={6000} 
+          onClose={()=>dispatch(setAlert({open: false, variant : null,massage : null}))}>
+                <Alert onClose={()=>dispatch(setAlert({open: false, variant : null,massage : null}))} severity = {alert.variant}  sx={{ width: '100%' }}>
+                    {alert.message}
                 </Alert>
             </Snackbar>
             }
