@@ -45,12 +45,12 @@ import {
 export default function Hinge() {
 
   const [search, setSearch] = useState("");
-  const [check,setCheck] = useState()
+  const [check,setCheck] = useState([])
 
 const {dispatch} = Store(); 
 const [pageSize, setPageSize] = useState(50);
 
-  const [Row, setRows] = useState()
+  const [Row, setRows] = useState([])
   // function for get cetegory list
 
   useEffect(() => {
