@@ -8,8 +8,8 @@ import {
 } from "@mui/material";
 import CreateIcon from '@mui/icons-material/Create';
 import AddIcon from "@mui/icons-material/Add";
-import { OpenBox, Notify } from "../../store/Types";
-import {Store } from "../../store/Context";
+import { useDispatch } from 'react-redux'
+import {setAlert,setForm } from "../../store/action/action";
 import { getPrimaryMaterial,changePrimaryMaterialStatus } from '../../services/service'
 import '../../assets/custom/css/category.css'
 
@@ -43,12 +43,12 @@ import {
 export default function PrimaryMaterial() {
 
   const [search, setSearch] = useState("");
-  const [check,setCheck] = useState()
+  const [check,setCheck] = useState([])
 
-const {dispatch} = Store(); 
+const dispatch = useDispatch(); 
 const [pageSize, setPageSize] = useState(50);
 
-  const [Row, setRows] = useState()
+  const [Row, setRows] = useState([])
   // function for get category list
 
   useEffect(() => {
@@ -116,13 +116,13 @@ const [pageSize, setPageSize] = useState(50);
       renderCell: (params) => 
       <div className="categoryImage" >
         <IconButton onClick={() => { 
-         dispatch({type : OpenBox,payload : {
+         dispatch(setForm({
             state : true,
             formType : 'update_PrimaryMaterial',
             payload : params,
             row : Row,
             setRow : setRows,
-          }}) 
+          })) 
         }} aria-label="delete"  >
           <CreateIcon />
         </IconButton>
@@ -160,21 +160,21 @@ const [pageSize, setPageSize] = useState(50);
         else 
         return row
       }))
-      dispatch({type : Notify,payload : {
+      dispatch(setAlert( {
         open : true,
         variant : 'success',
         message : " Material Status Updated Successfully !!!"
   
-      }})
+      }))
     })
     .catch((err)=>{
       //console.log(err)
-      dispatch({type : Notify,payload : {
+      dispatch(setAlert({
         open : true,
         variant : 'error',
         message : "Something Went Wrong !!!"
   
-      }})
+      }))
     })
 
     
@@ -246,7 +246,7 @@ const [pageSize, setPageSize] = useState(50);
         <Grid xs={12} md={2.8}>
           <Button
             onClick={() => {
-             dispatch({type : OpenBox,payload : { state: true, formType: "primaryMaterial",row : Row,setRow : setRows} });
+             dispatch(setForm({ state: true, formType: "primaryMaterial",row : Row,setRow : setRows} ));
             }}
             sx={{ width: "100%" }}
             color="primary"
