@@ -46,12 +46,12 @@ export default function Textile() {
 
 //   const [textile, setfabric] = useState("");
   const [search, setSearch] = useState("");
-  const [check,setCheck] = useState()
+  const [check,setCheck] = useState([])
 
 const {dispatch} = Store(); 
 const [pageSize, setPageSize] = useState(50);
 
-  const [Row, setRows] = useState()
+  const [Row, setRows] = useState([])
   // function for get category list
 
   useEffect(() => {
